requirements.txt
pandas
pyarrow
dash
plotly
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
